project/requirements.txt
Django>=5.2,<6
gunicorn==25.3.0
psycopg2-binary>=2.9
djangorestframework>=3.15
factory-boy>=3.3
coverage>=7.6
selenium>=4.20
webdriver-manager>=4.0
